requirements.txt
dspy>=2.4.0
groq>=0.4.1
python-dotenv>=1.0.0
